setfees seguir haciendo

+++ src/main/java/domain/fee.java
@@ -1,10 +1,10 @@
 package domain;
 
 public class fee {
-	private int fee;
+	private Float fee;
 	private String result;
 
-	public fee(int fee, String result) {
+	public fee(Float fee, String result) {
 		this.fee = fee;
 		this.result = result;
 	}
